픽어블 아이템 수정 백팩 기능 has empty space 구현 중 item info의 left top 으로 변경 및 int로 형변환
추가 작업 : hasEmptySpace에 리턴 타입을 튜플로 변경
addInventory 할때 invenVisualize에 = true 넣는 작업

그 이후 아이템 픽업 테스트

+++ Source/TarkovCopy/InventoryAndItem/ItemInfos/ItemInfo.h
@@ -19,12 +19,12 @@ public:
 	//virtual ~UItemInfo();
 
 	UPROPERTY(EditAnywhere)
-	float width;
+	int width;
 	UPROPERTY(EditAnywhere)
-	float height;
+	int height;
 
-	float left;
-	float top;
+	int left;
+	int top;
 
 	UPROPERTY(EditAnywhere)
 	int maxCapacity;
